Generalize the inversion function call.
The result is that changing the call to a routine that we know will
return the full inverse requires less than 10 lines when defining a
new class.

In the future, it may also

* read parameters from kwargs to make changing parameters in the
  inversion call trivial
* take a function directly from kwargs; this might be the better
  solution without adding complexity to the logic withing the class
  __init__ routines

Some predefined classes with fixed inv_funcs may be provided (say, the
difference pseudo-inverses).

Also removes some duplicate code.